--- a/native/com_wolfssl_WolfSSL.c
+++ b/native/com_wolfssl_WolfSSL.c
@@ -53,6 +53,16 @@ JavaVM*  g_vm;
 /* global object refs for logging callbacks */
 static jobject g_loggingCbIfaceObj;
 
+/* global method IDs we can cache for performance */
+jmethodID g_sslIORecvMethodId = NULL;
+jmethodID g_sslIOSendMethodId = NULL;
+jmethodID g_bufferPositionMethodId = NULL;
+jmethodID g_bufferLimitMethodId = NULL;
+jmethodID g_bufferHasArrayMethodId = NULL;
+jmethodID g_bufferArrayMethodId = NULL;
+jmethodID g_bufferSetPositionMethodId = NULL;
+jmethodID g_verifyCallbackMethodId = NULL;
+
 #ifdef HAVE_FIPS
 /* global object ref for FIPS error callback */
 static jobject g_fipsCbIfaceObj;
@@ -61,14 +71,114 @@ static jobject g_fipsCbIfaceObj;
 /* custom native fn prototypes */
 void NativeLoggingCallback(const int logLevel, const char *const logMessage);
 
-/* called when native library is loaded */
+/* Called when native library is loaded.
+ * We also cache global jmethodIDs here for performance. */
 JNIEXPORT jint JNICALL JNI_OnLoad(JavaVM* vm, void* reserved)
 {
+    JNIEnv* env = NULL;
+    jclass sslClass = NULL;
+    jclass byteBufferClass = NULL;
+    jclass verifyClass = NULL;
     (void)reserved;
 
     /* store JavaVM */
     g_vm = vm;
+
+    /* get JNIEnv from JavaVM */
+    if ((*vm)->GetEnv(vm, (void**)&env, JNI_VERSION_1_6) != JNI_OK) {
+        printf("Unable to get JNIEnv from JavaVM in JNI_OnLoad()\n");
+        return JNI_ERR;
+    }
+
+    /* Cache the method ID for IO send and recv callbacks */
+    sslClass = (*env)->FindClass(env, "com/wolfssl/WolfSSLSession");
+    if (sslClass == NULL) {
+        return JNI_ERR;
+    }
+
+    g_sslIORecvMethodId = (*env)->GetMethodID(env, sslClass,
+        "internalIOSSLRecvCallback",
+        "(Lcom/wolfssl/WolfSSLSession;[BI)I");
+
+    g_sslIOSendMethodId = (*env)->GetMethodID(env, sslClass,
+        "internalIOSSLSendCallback",
+        "(Lcom/wolfssl/WolfSSLSession;[BI)I");
+
+    /* Cache ByteBuffer method IDs */
+    byteBufferClass = (*env)->FindClass(env, "java/nio/ByteBuffer");
+    if (byteBufferClass == NULL) {
+        return JNI_ERR;
+    }
+
+    g_bufferPositionMethodId = (*env)->GetMethodID(env, byteBufferClass,
+        "position", "()I");
+    if (g_bufferPositionMethodId == NULL) {
+        return JNI_ERR;
+    }
+
+    g_bufferLimitMethodId = (*env)->GetMethodID(env, byteBufferClass,
+        "limit", "()I");
+    if (g_bufferLimitMethodId == NULL) {
+        return JNI_ERR;
+    }
+
+    g_bufferHasArrayMethodId = (*env)->GetMethodID(env, byteBufferClass,
+        "hasArray", "()Z");
+    if (g_bufferHasArrayMethodId == NULL) {
+        return JNI_ERR;
+    }
+
+    g_bufferArrayMethodId = (*env)->GetMethodID(env, byteBufferClass,
+        "array", "()[B");
+    if (g_bufferArrayMethodId == NULL) {
+        return JNI_ERR;
+    }
+
+    g_bufferSetPositionMethodId = (*env)->GetMethodID(env, byteBufferClass,
+        "position", "(I)Ljava/nio/Buffer;");
+    if (g_bufferSetPositionMethodId == NULL) {
+        return JNI_ERR;
+    }
+
+    /* Cache verify callback method ID */
+    verifyClass = (*env)->FindClass(env, "com/wolfssl/WolfSSLVerifyCallback");
+    if (verifyClass == NULL) {
+        return JNI_ERR;
+    }
+
+    g_verifyCallbackMethodId = (*env)->GetMethodID(env, verifyClass,
+        "verifyCallback", "(IJ)I");
+    if (g_verifyCallbackMethodId == NULL) {
+        return JNI_ERR;
+    }
+
+    /* Clean up local reference to class, not needed */
+    (*env)->DeleteLocalRef(env, sslClass);
+    (*env)->DeleteLocalRef(env, byteBufferClass);
+    (*env)->DeleteLocalRef(env, verifyClass);
+
     return JNI_VERSION_1_6;
+}
+
+/* Called when native library is unloaded.
+ * We clear cached method IDs here. */
+JNIEXPORT void JNICALL JNI_OnUnload(JavaVM* vm, void* reserved)
+{
+    JNIEnv* env;
+
+    if ((*vm)->GetEnv(vm, (void**)&env, JNI_VERSION_1_6) != JNI_OK) {
+        return;
+    }
+
+    /* Clear cached method ID */
+    g_sslIORecvMethodId = NULL;
+    g_sslIOSendMethodId = NULL;
+    g_bufferPositionMethodId = NULL;
+    g_bufferLimitMethodId = NULL;
+    g_bufferHasArrayMethodId = NULL;
+    g_bufferArrayMethodId = NULL;
+    g_bufferSetPositionMethodId = NULL;
+    g_verifyCallbackMethodId = NULL;
 }
 
 JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSL_init