This PR contains work done to increase `SSLEngine` (`WolfSSLEngine`) performance, including:

- Use a single static direct ByteBuffer in `WolfSSLEngine.SendAppData()`, instead of allocating a new array each time the method is called.
- Use a static output array for I/O output callback instead of allocating a new one for each send.
- Cache `jmethodID` values once globally when library is initialized, instead of multiple time throughout execution of code paths.
- Add new helper functions to throw exceptions (`throwWolfSSLJNIException()`, `throwWolfSSLException()`), which causes `FindClass()` to only be called when we need to throw an exception, not every time some JNI functions are called.
- Add ByteBuffer-based wolfSSL I/O callback support, and plug them into `WolfSSLEngine`. This allows our same native `char*` to be used again via wrapping with a direct ByteBuffer, instead of re-allocating another new array and copying data between Java and native C wolfSSL.

This also fixes runtime compatibility with the https://github.com/marianobarrios/tls-channel package:
- Adjusts `SSLSession.getPacketBufferSize()` for `tls-channel` compatibility. `tls-channel` uses a maximum I/O buffer size of 17k, and our `getPacketBufferSize()` was returning something too big.

Changes and performance were tested with a client and server demo that was created using `tls-channel` with wolfJSSE underneath. The client sent 500 MB to the server, the server sent the 500 MB back to the client, then the client compared the received data to the original sent.

Client-side benchmarks prior to these changes:

```
Client Benchmark Results:
Send: 524288000 bytes in 3.350 seconds (149.27 MB/s)
Receive: 524288000 bytes in 3.081 seconds (162.31 MB/s)
```

Client-side benchmarks after these changes show a **26% increase** in send and **31% increase** in receive throughput:

```
Client Benchmark Results:
Send: 524288000 bytes in 2.485 seconds (201.22 MB/s)
Receive: 524288000 bytes in 2.112 seconds (236.69 MB/s)
```